Handle multiple limit step with single measurement
If a Multiple*LimitStep only contains a single measurement
it will be considered as a Single*LimitStep when you upload
it. Adding a serialization hack that converts a
Multiple*LimitStep into its Single counterpart.

Removing the name of the measurement in the process,
since WATS doesn't allow for names for single
measurement steps.